Added filters and cleanup

+++ asciify.py
@@ -4,15 +4,50 @@ from tqdm import tqdm
 from moviepy.editor import VideoFileClip, AudioFileClip
 from moviepy.video.io.ffmpeg_writer import FFMPEG_VideoWriter
 from multiprocessing import Pool, cpu_count
-from PIL import Image, ImageDraw, ImageFont
+from PIL import Image, ImageDraw, ImageFont, ImageFilter
 import numpy as np
 import os
 
+CRT_SCAN_LINE_COLOR = 50
+
+def apply_crt_effect(image):
+    crt_image = image.copy()
+    width, height = crt_image.size
+    pixels = crt_image.load()
+    
+    for y in range(height):
+        if y % 2 == 0:
+            for x in range(width):
+                r, g, b = pixels[x, y]
+                pixels[x, y] = (r // 2, g // 2, b // 2)
+    
+    crt_image = crt_image.filter(ImageFilter.GaussianBlur(1))
+    return crt_image
+
+def apply_sepia_filter(image):
+    sepia_image = np.array(image)
+    sepia_filter = np.array([[0.272, 0.534, 0.131],
+                             [0.349, 0.686, 0.168],
+                             [0.393, 0.769, 0.189]])
+                             
+    sepia_image = cv2.transform(sepia_image, sepia_filter)
+    sepia_image = np.clip(sepia_image, 0, 255)
+    return Image.fromarray(sepia_image.astype(np.uint8))
+
+def apply_color_tint(image, tint_color):
+    tinted_image = Image.new("RGB", image.size)
+    image = image.convert("RGB") 
+    tint_color_image = Image.new("RGB", tinted_image.size, tint_color)
+    blended = Image.blend(image, tint_color_image, alpha=0.3)
+    return blended
+
 def asciify_frame(
-        frame, scaleFactor=0.15, oneCharWidth=7, oneCharHeight=9, color_brightness=1, pixel_brightness=2.15, monochrome=False, 
+        frame, scale_factor=0.15, one_char_width=8, one_char_height=8, color_brightness=1, pixel_brightness=2.15, monochrome=False, filters=None,
         overlay_contours=False, contour_depth_minimum_threshold = 0, contour_depth_maximum_threshold = 255, progress_bar=False
     ):
-    chars = "$@B%8&WM#*oahkbdpqwmZO0QLCJUYXzcvunxrjft/\\|()1{}[]?-_+~<>i!lI;:,\"^`'. "[::-1]
+    
+    chars = "■@?0Poc;. "[::-1]
+    # chars = "$@B%8&WM#*oahkbdpqwmZO0QLCJUYXzcvunxrjft/\\|()1{}[]?-_+~<>i!lI;:,\"^`'. "[::-1]
     charArray = list(chars)
     charLength = len(charArray)
     interval = charLength / 256
@@ -22,19 +57,20 @@ def asciify_frame(
     def getChar(inputInt):
         return charArray[math.floor(inputInt * interval)]
     
-    if frame.shape[2] == 4:
-        frame = frame[:, :, :3]
-    frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
+    if frame.shape[2] == 3:
+        frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
+    elif frame.shape[2] == 4:
+        frame_rgb = cv2.cvtColor(frame[:, :, :3], cv2.COLOR_BGR2RGB)
     
     im = Image.fromarray(frame_rgb)
     fnt = ImageFont.truetype('lucon.ttf', 10)
 
     width, height = im.size
-    im = im.resize((int(scaleFactor * width), int(scaleFactor * height * (oneCharWidth / oneCharHeight))), Image.NEAREST)
+    im = im.resize((int(scale_factor * width), int(scale_factor * height * (one_char_width / one_char_height))), Image.NEAREST)
     width, height = im.size
     pix = im.load()
 
-    outputImage = Image.new('RGB', (int(oneCharWidth * width), int(oneCharHeight * height)), color=(0, 0, 0))
+    outputImage = Image.new('RGB', (int(one_char_width * width), int(one_char_height * height)), color=(0, 0, 0))
     d = ImageDraw.Draw(outputImage)
 
     gray = cv2.cvtColor(np.array(im), cv2.COLOR_BGR2GRAY)
@@ -77,9 +113,18 @@ def asciify_frame(
             if monochrome:
                 rgb = (h,h,h)
 
-            d.text((j * oneCharWidth, i * oneCharHeight), char, font=fnt, fill=rgb, align="center")
+            d.text((j * one_char_width, i * one_char_height), char, font=fnt, fill=rgb, align="center")
     
     outputImage = outputImage.convert('RGB')
+    
+    if filters:
+        if 'crt' in filters:
+            outputImage = apply_crt_effect(outputImage)
+        if 'sepia' in filters:
+            outputImage = apply_sepia_filter(outputImage)
+        if 'tint' in filters and 'tint_color' in filters:
+            outputImage = apply_color_tint(outputImage, filters['tint_color'])
+    
     output_frame = np.array(outputImage)
     return output_frame
 
@@ -87,7 +132,7 @@ def convert_frame(args):
     return asciify_frame(*args)
 
 def ascii_photo(
-        in_path, final_path, scaleFactor=0.15, oneCharWidth=7, oneCharHeight=9, color_brightness=1, pixel_brightness=2.15, monochrome=False, 
+        in_path, final_path, scale_factor=0.15, one_char_width=8, one_char_height=8, color_brightness=1, pixel_brightness=2.15, monochrome=False, filters=None,
         overlay_contours=False, contour_depth_minimum_threshold = 0, contour_depth_maximum_threshold = 255, progress_bar=False
     ):
     """
@@ -101,12 +146,12 @@ def ascii_photo(
     - scaleFactor: Float.
         Controls the image quality in the ASCII image.
         Default is 0.15.
-    - oneCharWidth: Int.
+    - one_char_width: Int.
         Width of one character in the ASCII representation.
-        Default is 7.
-    - oneCharHeight: Int.
+        Default is 8.
+    - one_char_height: Int.
         Height of one character in the ASCII representation.
-        Default is 9.
+        Default is 8.
     - color_brightness: Float.
         Specify brightness multiplier of colors in calculations.
         Default is 1.
@@ -116,6 +161,9 @@ def ascii_photo(
     - monochrome: Bool.
         If True, a frames are rendered using only grayscale colors.
         Default is False.
+    - filters: Dict.
+        A dictionary containing filters to use. crt, sepia, and tint are boolean keys, and tint requires a tint_color key with a color tuple (0-255).
+        Default is None.
     - overlay_contours: Bool.
         If True, overlays contours on the image.
         Default is False.
@@ -138,7 +186,7 @@ def ascii_photo(
 
     cap = cv2.imread(in_path)
     frame_args = (
-        cap, scaleFactor, oneCharWidth, oneCharHeight, color_brightness, pixel_brightness, monochrome, 
+        cap, scale_factor, one_char_width, one_char_height, color_brightness, pixel_brightness, monochrome, filters,
         overlay_contours, contour_depth_minimum_threshold, contour_depth_maximum_threshold, progress_bar
     )
     ascii_frame = Image.fromarray(convert_frame(frame_args), 'RGB')
@@ -148,7 +196,7 @@ def ascii_photo(
 
 
 def ascii_video(
-        in_path, final_path, scaleFactor = 0.15, oneCharWidth = 7, oneCharHeight = 9, color_brightness=1, pixel_brightness=2.15, monochrome=False, 
+        in_path, final_path, scale_factor = 0.15, one_char_width = 8, one_char_height = 8, color_brightness=1, pixel_brightness=2.15, monochrome=False, filters=None,
         overlay_contours = False, contour_depth_minimum_threshold = 0, contour_depth_maximum_threshold = 255, low_res_audio = True, progress_bar = True, num_workers=None
     ):
     """
@@ -159,15 +207,15 @@ def ascii_video(
         Relative path to the input video.
     - final_path: String.
         Relative path to the output video. Doesn't need to exist.
-    - scaleFactor: Float.
+    - scale_factor: Float.
         Controls the video quality in the ASCII video.
         Default is 0.15.
-    - oneCharWidth: Int.
+    - one_char_width: Int.
         Width of one character in the ASCII representation.
-        Default is 7.
-    - oneCharHeight: Int.
+        Default is 8.
+    - one_char_height: Int.
         Height of one character in the ASCII representation.
-        Default is 9.
+        Default is 8.
     - color_brightness: Float.
         Specify brightness multiplier of colors in calculations.
         Default is 1.
@@ -177,6 +225,9 @@ def ascii_video(
     - monochrome: Bool.
         If True, a frames are rendered using only grayscale colors.
         Default is False.
+    - filters: Dict.
+        A dictionary containing filters to use. crt, sepia, and tint are boolean keys, and tint requires a tint_color key with a color tuple (0-255).
+        Default is None.
     - overlay_contours: Bool.
         If True, overlays contours on the video.
         Default is False.
@@ -221,7 +272,7 @@ def ascii_video(
         ret, frame = cap.read()
         if ret:
             frame_args.append((
-                frame, scaleFactor, oneCharWidth, oneCharHeight, color_brightness, pixel_brightness, monochrome, 
+                frame, scale_factor, one_char_width, one_char_height, color_brightness, pixel_brightness, monochrome, filters,
                 overlay_contours, contour_depth_minimum_threshold, contour_depth_maximum_threshold, False
             ))
         else:
@@ -239,7 +290,7 @@ def ascii_video(
     output_video = cv2.VideoWriter(temp_video_path, fourcc, frame_rate, (ascii_frames[0].shape[1], ascii_frames[0].shape[0]))
 
     for frame in tqdm(ascii_frames, desc="Saving video", disable = not progress_bar, leave=False):
-        output_video.write(frame)
+        output_video.write(cv2.cvtColor(frame, cv2.COLOR_RGB2BGR))
     
     output_video.release()
 
@@ -253,15 +304,18 @@ def ascii_video(
     temp_audio_path = "temp_audio.mp3"
     temp_audio_low_path = "temp_audio_low.mp3"
 
-    if low_res_audio:
-        low_audio = original_audio.set_fps(8000)
-        low_audio.write_audiofile(temp_audio_low_path, verbose=False, logger=None)
-        
-        audio = AudioFileClip(temp_audio_low_path)
-        audio = audio.set_fps(16000)
-        audio.write_audiofile(temp_audio_path, verbose=False, logger=None)
+    if original_audio != None:
+        if low_res_audio:
+            low_audio = original_audio.set_fps(8000)
+            low_audio.write_audiofile(temp_audio_low_path, verbose=False, logger=None)
+            
+            audio = AudioFileClip(temp_audio_low_path)
+            audio = audio.set_fps(16000)
+            audio.write_audiofile(temp_audio_path, verbose=False, logger=None)
+        else:
+            original_audio.write_audiofile(temp_audio_path, verbose=False, logger=None)
     else:
-        original_audio.write_audiofile(temp_audio_path, verbose=False, logger=None)
+        temp_audio_path = None
 
     with FFMPEG_VideoWriter(final_path, fps=fps, size=clip.size, codec='libx264', logfile=None, threads=num_workers, audiofile=temp_audio_path, ffmpeg_params=['-strict', '-2']) as writer:
         frame_iterator = tqdm(clip.iter_frames(fps), total=int(clip.duration*fps), desc = "Adding audio", disable = not progress_bar, leave=False)
@@ -274,5 +328,6 @@ def ascii_video(
     clip.close()
     original_clip.close()
     os.remove(temp_video_path)
-    if low_res_audio: os.remove(temp_audio_low_path)
-    os.remove(temp_audio_path)
+    if temp_audio_path:
+        if low_res_audio: os.remove(temp_audio_low_path)
+        os.remove(temp_audio_path)
